Enum pallet_plasm_lockdrop::Call

pub enum Call<T: Trait> {
    request(Lockdrop, H256),
    claim(ClaimId),
    claim_to(ClaimId, T::AccountId, EcdsaSignature),
    vote(ClaimVote, <T::AuthorityId as RuntimeAppPublic>::Signature),
    set_dollar_rate(TickerRate<T::DollarRate>, <T::AuthorityId as RuntimeAppPublic>::Signature),
    set_alpha(u32),
    set_bounds(T::BlockNumber, T::BlockNumber),
    set_positive_votes(AuthorityVote),
    set_vote_threshold(AuthorityVote),
    set_authorities(Vec<T::AuthorityId>),
    // some variants omitted
}

Dispatchable calls.

Each variant of this enum maps to a dispatchable function from the associated module.

Variants

request(Lockdrop, H256)

Request authorities to check locking transaction. TODO: weight

claim(ClaimId)

Claim tokens for the lockdrop public key. TODO: weight

claim_to(ClaimId, T::AccountId, EcdsaSignature)

Claim tokens for any account address. TODO: weight

vote(ClaimVote, <T::AuthorityId as RuntimeAppPublic>::Signature)

Vote for claim request according to check results. (for authorities only) TODO: weight

set_dollar_rate(TickerRate<T::DollarRate>, <T::AuthorityId as RuntimeAppPublic>::Signature)

Dollar Rate oracle entrypoint. (for authorities only) TODO: weight

set_alpha(u32)

Set lockdrop alpha value.

set_bounds(T::BlockNumber, T::BlockNumber)

Set lockdrop held time.

set_positive_votes(AuthorityVote)

Set minimum of positive votes required for lock approve.

set_vote_threshold(AuthorityVote)

Set minimum votes required to pass lock confirmation process.

set_authorities(Vec<T::AuthorityId>)

Set lockdrop authorities list.