Struct pallet_plasm_lockdrop::Module

pub struct Module<T: Trait>(_);

Implementations

impl<T: Trait + 'static> Module<T>

pub fn requests() -> Vec<ClaimId>

Offchain lock check requests made within this block execution.

pub fn keys() -> Vec<T::AuthorityId>

List of lockdrop authority id's.

pub fn claims<K: EncodeLike<ClaimId>>(key: K) -> Claim<T::AuthorityId>

Token claim requests.

pub fn has_vote<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> bool where
KArg1: EncodeLike<T::AuthorityId>,
KArg2: EncodeLike<ClaimId>,

pub fn alpha() -> Perbill

Lockdrop alpha parameter, where α ∈ [0; 1]

pub fn dollar_rate() -> (T::DollarRate, T::DollarRate)

Lockdrop dollar rate parameter: BTC, ETH.

pub fn dollar_rate_f<K: EncodeLike<T::AuthorityId>>(
key: K
) -> (T::Moment, T::DollarRate, T::DollarRate)

Lockdrop dollar rate median filter table: Time, BTC, ETH.

pub fn vote_threshold() -> AuthorityVote

How much authority votes module should receive to decide claim result.

pub fn positive_votes() -> AuthorityVote

How much positive votes requered to approve claim. Positive votes = approve votes - decline votes.

pub fn lockdrop_bounds() -> (T::BlockNumber, T::BlockNumber)

Timestamp bounds of lockdrop held period.